Paso 51

+++ database/migrations/2023_12_07_154300_add_team_id_to_users.php
@@ -14,7 +14,7 @@ class AddTeamIdToUsers extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedInteger('team_id')
+            $table->unsignedBigInteger('team_id')
                 ->nullable()
                 ->after('id');
